Stopping Criterion

Stopping Criterion

StochOptInterface.stop — Method.

stop(s::AbstractStoppingCriterion, info::Info)

Determine whether the SDDP algorithm should stop using the information stored in info.

source

StochOptInterface.OrStoppingCriterion — Type.

type OrStoppingCriterion <: StochOptInterface.AbstractStoppingCriterion

Stops if lhs or rhs want to stop.

source

StochOptInterface.AndStoppingCriterion — Type.

type AndStoppingCriterion <: StochOptInterface.AbstractStoppingCriterion

Stops if lhs and rhs want to stop.

source

StochOptInterface.IterLimit — Type.

type IterLimit <: StochOptInterface.AbstractStoppingCriterion

Stops if iter ≧ limit.

source

StochOptInterface.CutLimit — Type.

type CutLimit <: StochOptInterface.AbstractStoppingCriterion

Stops if there was less than or equal to limit cuts added in the iteration. For instance, CutLimit(0) stops when there are no cuts added.

source

StochOptInterface.Pereira — Type.

type Pereira <: StochOptInterface.AbstractStoppingCriterion

Stops if z_UB - α * σ/√K - tol < z_LB < z_UB + α * σ/√K + tol and σ / √K > β * max(1, |z_LB|))

source