Stopping Criterion

StochOptInterface.stop — Method

stop(s::AbstractStoppingCriterion, info::Info)

Determine whether the SDDP algorithm should stop using the information stored in info.

StochOptInterface.OrStoppingCriterion — Type

mutable struct OrStoppingCriterion <: StochOptInterface.AbstractStoppingCriterion

Stop if lhs or rhs want to stop.

StochOptInterface.AndStoppingCriterion — Type

mutable struct AndStoppingCriterion <: StochOptInterface.AbstractStoppingCriterion

Stop if lhs and rhs want to stop.

StochOptInterface.IterLimit — Type

mutable struct IterLimit <: StochOptInterface.AbstractStoppingCriterion

Stop if iter ≧ limit.

StochOptInterface.CutLimit — Type

mutable struct CutLimit <: StochOptInterface.AbstractStoppingCriterion

Stop if there was less than or equal to limit cuts added in the iteration. For instance, CutLimit(0) stops when there are no cuts added.

StochOptInterface.Pereira — Type

mutable struct Pereira <: StochOptInterface.AbstractStoppingCriterion

Stop if z_UB - α * σ/√K - tol < z_LB < z_UB + α * σ/√K + tol and σ / √K > β * max(1, |z_LB|))